requirements.txt
PyPDF2
pinecone
python-dotenv